Remove check_mod_loops query and run the checks per-body instead
