+++ compiler/rustc_hir_typeck/src/lib.rs
@@ -28,6 +28,7 @@ mod fallback;
 mod fn_ctxt;
 mod gather_locals;
 mod intrinsicck;
+mod loops;
 mod method;
 mod naked_functions;
 mod op;
@@ -149,6 +150,7 @@ fn typeck_with_inspect<'tcx>(
         };
 
         check_abi(tcx, id, span, fn_sig.abi());
+        loops::check(tcx, def_id, body);
 
         // Compute the function signature from point of view of inside the fn.
         let mut fn_sig = tcx.liberate_late_bound_regions(def_id.to_def_id(), fn_sig);
@@ -188,6 +190,13 @@ fn typeck_with_inspect<'tcx>(
         } else {
             tcx.type_of(def_id).instantiate_identity()
         };
+
+        // TODO: anon consts are currently loop checked with their containing body, even though
+        // they are typecked on their own.
+        if let DefKind::AssocConst | DefKind::Const | DefKind::Static { .. } = tcx.def_kind(def_id)
+        {
+            loops::check(tcx, def_id, body);
+        }
 
         let expected_type = fcx.normalize(body.value.span, expected_type);
